requirements.txt
opencv-python
pyttsx3
pytesseract
pillow
